+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/gui/NewPlanDialog.java
@@ -40,7 +40,6 @@ import org.zaproxy.addon.automation.AutomationJob;
 import org.zaproxy.addon.automation.AutomationPlan;
 import org.zaproxy.addon.automation.ExtensionAutomation;
 import org.zaproxy.addon.automation.jobs.ActiveScanJob;
-import org.zaproxy.addon.automation.jobs.AddOnJob;
 import org.zaproxy.addon.automation.jobs.PassiveScanConfigJob;
 import org.zaproxy.addon.automation.jobs.PassiveScanWaitJob;
 import org.zaproxy.addon.automation.jobs.SpiderJob;
@@ -68,7 +67,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
     private static final String REPORT_JOB_NAME = "report";
 
     private static final String[] BASELINE_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         SpiderJob.JOB_NAME,
         "spiderAjax",
@@ -76,7 +74,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
         REPORT_JOB_NAME
     };
     private static final String[] IMPORT_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         "import",
         SpiderJob.JOB_NAME,
@@ -86,7 +83,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
         REPORT_JOB_NAME
     };
     private static final String[] OPENAPI_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         "openapi",
         PassiveScanWaitJob.JOB_NAME,
@@ -94,7 +90,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
         REPORT_JOB_NAME
     };
     private static final String[] GRAPHQL_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         "graphql",
         PassiveScanWaitJob.JOB_NAME,
@@ -102,7 +97,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
         REPORT_JOB_NAME
     };
     private static final String[] SOAP_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         "soap",
         PassiveScanWaitJob.JOB_NAME,
@@ -110,7 +104,6 @@ public class NewPlanDialog extends StandardFieldsDialog {
         REPORT_JOB_NAME
     };
     private static final String[] FULL_SCAN_PROFILE = {
-        AddOnJob.JOB_NAME,
         PassiveScanConfigJob.JOB_NAME,
         SpiderJob.JOB_NAME,
         "spiderAjax",
@@ -146,6 +139,7 @@ public class NewPlanDialog extends StandardFieldsDialog {
         List<AutomationJob> jobs =
                 ext.getAutomationJobs().values().stream()
                         .filter(j -> !j.isDataJob())
+                        .filter(j -> j.getClass().getAnnotation(Deprecated.class) == null)
                         .collect(Collectors.toList());
 
         Collections.sort(jobs);