+++ addOns/automation/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Maintenance changes.
 - Rely on spider add-on (Issue 3113).
 - passiveScan-config job resets the state, as scanOnlyInScope is often confusing in the GUI. 
+- Deprecated the addOns job.
 
 ### Fixed
 - Correct loading of custom scripts (e.g. Zest).